Create L-layer.JPG
The initialization for a deeper L-layer neural network is more complicated because there are many more weight matrices and bias vectors. When completing the initialize_parameters_deep function, make sure that the dimensions match between each layer. Recall that  𝑛[𝑙]  is the number of units in layer  𝑙 . For example, if the size of your input  𝑋  is  (12288,209)  (with  𝑚=209  examples)